requirements.txt
pandas
imdbpy
selenium
webdriver-manager
